chore: Expose more bindings (#13)
* re-add msgpackr to runtime

JSON Payloads are hard to deal with in Protobuf since we really can't tell the structure of the JSON forehand. And this, msgpackr handles very well between
Dart & Javascript/Typescript conversions.

* fix msgpackr in runtime

* fix formatting

* _

* expose JsonPayload interface instead of MessagePackr

This is so we don't expose too much implementation details, making it easier to swap msgpackr later in the future
if it doesn't hold up well.

* handle null values passed from Dart